+++ app/src/main/java/com/runwise/supply/orderpage/entity/ProductData.java
@@ -38,7 +38,7 @@ public class ProductData implements Serializable{
 
         private double actualQty;
         private boolean isTwoUnit;
-        private int presetQty;
+        private double presetQty;
         private String settlePrice;
         private int productID;
         private String priceID;
@@ -133,11 +133,11 @@ public class ProductData implements Serializable{
             this.isTwoUnit = isTwoUnit;
         }
 
-        public int getPresetQty() {
+        public double getPresetQty() {
             return presetQty;
         }
 
-        public void setPresetQty(int presetQty) {
+        public void setPresetQty(double presetQty) {
             this.presetQty = presetQty;
         }
 
@@ -269,7 +269,7 @@ public class ProductData implements Serializable{
         public void writeToParcel(Parcel dest, int flags) {
             dest.writeDouble(this.actualQty);
             dest.writeByte(this.isTwoUnit ? (byte) 1 : (byte) 0);
-            dest.writeInt(this.presetQty);
+            dest.writeDouble(this.presetQty);
             dest.writeString(this.settlePrice);
             dest.writeInt(this.productID);
             dest.writeString(this.priceID);
@@ -296,7 +296,7 @@ public class ProductData implements Serializable{
         protected ListBean(Parcel in) {
             this.actualQty = in.readDouble();
             this.isTwoUnit = in.readByte() != 0;
-            this.presetQty = in.readInt();
+            this.presetQty = in.readDouble();
             this.settlePrice = in.readString();
             this.productID = in.readInt();
             this.priceID = in.readString();
